agent/requirements.txt
langchain==0.3.27
langgraph==0.6.6
langsmith==0.4.23
openai>=1.68.2,<2.0.0
fastapi>=0.115.5,<1.0.0
uvicorn>=0.29.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
langgraph-cli[inmem]==0.3.3
langchain-openai>=0.0.1
tavily-python>=0.0.0
redis[async]>=0.0.0
langcache>=0.0.0